requirements.txt
et-xmlfile>=1.1.0
openpyxl>=3.0.9
PyYAML==6.0
